Going to merge this as the original intent has been achieved. My realest estimate on code coverage with this is maybe 35-40%, but I plan to work in actual coverage metrics soon. I ran the ::NG_Stress() static variant today and it took nearly 5 minutes but passed 100% of ~5700 CRUD test cases (single thread still) with most notable performance hits in user instantiation which I presume is due to its cryptographically heavy nature.

The next priority is ensure all command handlers are checking all aliases (as in /protocols/shell/commands/role.cpp in this merge) so that the command handling becomes 100% driven by the CLI CommandUsage definitions.

Its only onward and upward from here.

+++ src/types/S3Vault.cpp
@@ -17,7 +17,7 @@ S3Vault::S3Vault(const std::string& name, const unsigned int apiKeyID, std::stri
 
 S3Vault::S3Vault(const pqxx::row& row)
     : Vault(row),
-      api_key_id(row["api_key_id"].as<unsigned short>()),
+      api_key_id(row["api_key_id"].as<unsigned int>()),
       bucket(row["bucket"].as<std::string>()),
       encrypt_upstream(row["encrypt_upstream"].as<bool>()) {}
 
@@ -30,7 +30,7 @@ void to_json(nlohmann::json& j, const S3Vault& v) {
 
 void from_json(const nlohmann::json& j, S3Vault& v) {
     from_json(j, static_cast<Vault&>(v));
-    v.api_key_id = j.at("api_key_id").get<unsigned short>();
+    v.api_key_id = j.at("api_key_id").get<unsigned int>();
     v.bucket = j.at("bucket").get<std::string>();
     v.encrypt_upstream = j.value("encrypt_upstream", true);
 }
